requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.0
rasterio>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
wandb>=0.15.0
shapely>=2.0.0
opencv-python>=4.7.0
pyyaml>=6.0
scikit-learn>=1.2.0
tqdm>=4.65.0
